refactor(cockpit): extract proof evidence normalization
